add oneway collider on door, attack of player2 on right ctrl, add delta time for ghost attack

+++ Game/Assets/Scripts/Health/HealthBarForPlayer.cs
@@ -12,11 +12,11 @@ public class HealthBarForPlayer : MonoBehaviour
 
     private void Start()
     {
-        totalHealthBar.fillAmount = playerHealth.currentHealth / 20;
+        totalHealthBar.fillAmount = (float)(playerHealth.currentHealth / 20);
     }
 
     private void Update()
     {
-        currentHealthBar.fillAmount = playerHealth.currentHealth / 20;
+        currentHealthBar.fillAmount = (float)(playerHealth.currentHealth / 20);
     }
 }
